Add activity_failed_logger
